requirements.txt
install-ifcopenshell-python==0.1.0
numpy==1.24.3
pyproj==3.5.0

